py185/my_films/pyproject.toml
[tool.poetry]
name = "my-films"
version = "0.1.0"
description = ""
authors = ["Deepak Karivelil <deepakkarivelil@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
